+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - Method to check if drive is connected via a COM-KIT.
+- Property to check if the PDOs are active.
 
 ## [0.8.2] - 2024-07-15
 ### Added