+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [0.8.3] - 2024-07-26
 ### Changed
 - Make fsoe_master an optional requirement